修复bug，以及增加查询数据接口；

+++ core/src/main/java/com/mingshi/skyflying/domain/UserPortraitByVisitedTableEverydayDo.java
@@ -7,7 +7,7 @@ import java.util.Date;
 public class UserPortraitByVisitedTableEverydayDo {
     private Integer id;
 
-    private Byte isDelete;
+    private Integer isDelete;
 
     private Date gmtCreate;
